chore: New folder structure
I made the package `expressions` for expression types, and all the logic went to the main package.
