--- src/main/scala/replcalc/eval/Expression.scala
@@ -1,5 +1,0 @@
-package replcalc.eval
-
-trait Expression:
-  def evaluate: Either[Error, Double]
-
