+++ src/main/scala/replcalc/expressions/Constant.scala
@@ -1,6 +1,7 @@
-package replcalc.eval
+package replcalc.expressions
 
 import Error.ParsingError
+import replcalc.Parser
 
 final case class Constant(number: Double) extends Expression:
   override def evaluate: Either[Error, Double] = Right(number)
