+++ src/main/scala/replcalc/expressions/Value.scala
@@ -1,6 +1,7 @@
-package replcalc.eval
+package replcalc.expressions
 
 import Error.ParsingError
+import replcalc.Parser
 
 final case class Value(name: String, expr: Expression) extends Expression:
   override def evaluate: Either[Error, Double] = expr.evaluate
