+++ src/test/scala/replcalc/PreprocessorTest.scala
@@ -1,4 +1,4 @@
-package replcalc.eval
+package replcalc
 
 import munit.Location
